Fixed misc. bugs, including list functionality (select still broken).

+++ core/src/main/java/net/jselby/escapists/data/chunks/Assets.kt
@@ -329,9 +329,11 @@ class SoundBank : Chunk() {
             if (!directoryFile.exists() && !directoryFile.mkdirs()) {
                 throw IllegalStateException("Failed to create audio directory.");
             }
-            val out = FileOutputStream("$directory$name.$extension");
-            out.write(data.getBytes(decompressedLength));
-            out.close();
+            if (!File("$directory$name.$extension").exists()) {
+                val out = FileOutputStream("$directory$name.$extension");
+                out.write(data.getBytes(decompressedLength));
+                out.close();
+            }
         }
     }
 }
